feat: 요구사항 2(구입금액에 만큼 로또 번호 자동으로 생성한다)
- 로또 자동 생성 및 출력

+++ src/main/kotlin/domain/Amount.kt
@@ -1,8 +1,8 @@
 package domain
 
-const val LOTTO_PRICE = 1000L
+const val LOTTO_PRICE = 1000
 
-data class Amount(val amount: Long) {
+data class Amount(val amount: Int) {
 
     fun calculatePurchaseLottoCount() = amount / LOTTO_PRICE
